refactor: 경로 포맷터 공유 및 prop 기본값 정리
## 📌 작업 내용
- 경로 거리/시간 포맷터를 공용 유틸로 추출
- 추천 관련 컴포넌트의 함수형 prop 기본값을 빈 함수로 통일
- 홈 지도 트러블슈팅 문서 업데이트 반영

## 📁 변경된 파일
- frontend/src/utils/formatters.js
- frontend/src/views/HomeView.vue
- frontend/src/views/restaurant/id/RestaurantDetailPage.vue
- frontend/src/components/ui/RestaurantCardList.vue
- frontend/src/components/ui/HomeRecommendationContent.vue
- frontend/src/components/ui/CafeteriaRecommendationSection.vue
- frontend/src/components/ui/TrendingRecommendationSection.vue
- docs/home-map-troubleshooting.md

## 공유사항 to 리뷰어
- 포맷터 유틸 공유로 중복 제거, 추천 컴포넌트 prop 기본값 변경 확인 부탁

## 🔗 관련 Issue(선택)
-

## ✔️ 체크리스트(선택)
-
