+++ frontend/src/components/ui/HomeRecommendationContent.vue
@@ -86,7 +86,7 @@ defineProps({
   },
   onCheckRoute: {
     type: Function,
-    default: null,
+    default: () => {},
   },
   routeLoadingId: {
     type: [Number, String],
