+++ frontend/src/components/ui/RestaurantCardList.vue
@@ -3,6 +3,10 @@ import { RouterLink } from "vue-router";
 import { Star } from "lucide-vue-next";
 import Card from "@/components/ui/Card.vue";
 import FavoriteStarButton from "@/components/ui/FavoriteStarButton.vue";
+import {
+  formatRouteDistance,
+  formatRouteDurationMinutes,
+} from "@/utils/formatters";
 
 const props = defineProps({
   restaurants: {
@@ -19,7 +23,7 @@ const props = defineProps({
   },
   onCheckRoute: {
     type: Function,
-    default: null,
+    default: () => {},
   },
   routeLoadingId: {
     type: [Number, String],
@@ -93,17 +97,7 @@ const isRouteActive = (restaurant) => {
   return String(currentId) === String(activeId);
 };
 
-const formatRouteDistance = (meters) => {
-  if (!Number.isFinite(meters)) return "-";
-  if (meters < 1000) return `${Math.round(meters)}m`;
-  return `${(meters / 1000).toFixed(1)}km`;
-};
-
-const formatRouteDuration = (seconds) => {
-  if (!Number.isFinite(seconds)) return "-";
-  const totalMinutes = Math.max(1, Math.round(seconds / 60));
-  return `${totalMinutes}분`;
-};
+const formatRouteDuration = formatRouteDurationMinutes;
 </script>
 
 <template>
