+++ frontend/src/components/ui/TrendingRecommendationSection.vue
@@ -30,7 +30,7 @@ defineProps({
   },
   onCheckRoute: {
     type: Function,
-    default: null,
+    default: () => {},
   },
   routeLoadingId: {
     type: [Number, String],
